[REFACTOR] 프로덕트 도메인 인기도 처리 로직 추가 및 리팩토링-S09P22A204-529

See merge request s09-bigdata-recom-sub2/S09P22A204!98

+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/product/domain/CoffeeReview.java
@@ -8,8 +8,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
-import javax.persistence.*;
-import java.time.LocalDateTime;
+import javax.persistence.AttributeOverride;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 
 @Getter
 @SuperBuilder
@@ -31,9 +35,9 @@ public class CoffeeReview extends BaseEntity {
     private String content;
 
     @Column
-    private Double score;
+    private Integer score;
 
-    public void update(String content, Double score) {
+    public void update(String content, Integer score) {
         this.content = content;
         this.score = score;
     }
