+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/product/dto/ProductReviewElement.java
@@ -2,7 +2,7 @@ package com.ssafy.coffeeing.modules.product.dto;
 
 public record ProductReviewElement(
         Long reviewId,
-        Double score,
+        Integer score,
         String content,
         Long memberId,
         String profileImageUrl,
